Webserive view binding done, working on offline fetch

+++ app/src/main/java/com/gadgeon_mt/weatherforecast/database/WeatherListDb.kt
@@ -3,6 +3,6 @@ package com.gadgeon_mt.weatherforecast.database
 data class WeatherListDb (
     val temp_min: Double,
     val temp_max: Double,
-    val humidity: Double,
+    val humidity: Int,
     val dateTime: String,
 )
